requirements.txt
certifi==2022.9.24
click==8.1.3
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-amazon-paapi==5.0.0
python-dateutil==2.8.2
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0
